security: comprehensive production readiness hardening (12-round audit)
Systematic audit→remediate→re-audit loop (12 rounds) covering security,
data integrity, error handling, infrastructure, and frontend quality.

Security:
- Remove hardcoded JWT fallback, fail-closed auth/CORS in production
- HTTP server timeouts (Read/Write/Idle), panic recovery middleware
- RBAC applied to all 30+ handler modules via RequireRole middleware
- Rate limiting (global + strict login), request size limits (10MB)
- Sanitize all error responses (httputil.RespondError), no internal leaks
- Content-Disposition: attachment on uploads, security headers in nginx
- Portal ListUsers admin-gated, integration auth constant-time compare

Data integrity:
- Financial ops wrapped in RunInTx (invoice, credit memo, AP, POS, orders)
- FOR UPDATE on customer balance reads (no lost-update race)
- Payment/refund/credit-memo amount validation (positive only)
- AP GL sync errors now roll back transactions (no silent desync)
- Credit memo create+apply atomic (CreateAndApplyCreditMemo)
- POS line total uses math.Round (no float truncation)

Frontend:
- fetchWithAuth wrapper with timeout/retry on all API services
- Error boundaries on route groups, toast notifications on all errors
- All catch(err: any)→catch(err: unknown), all as-any eliminated
- setTimeout/setInterval stored in refs with cleanup useEffects
- useEffect/useCallback dependency arrays corrected across ~30 files
- Blob URL revocation, promise .catch() handlers on all chains

Infrastructure:
- Docker non-root users, healthchecks, build flags (-ldflags -trimpath)
- CI pipeline: test, lint, build, govulncheck, permissions, concurrency
- Audit logger with WaitGroup drain on graceful shutdown
- docker-compose ports bound to 127.0.0.1

+++ backend/internal/order/model.go
@@ -22,7 +22,7 @@ type Order struct {
 	CustomerName string      `json:"customer_name,omitempty"`
 	QuoteID      *uuid.UUID  `json:"quote_id,omitempty"`
 	Status       OrderStatus `json:"status"`
-	TotalAmount  float64     `json:"total_amount"`
+	TotalAmount  int64       `json:"total_amount"` // Cents
 	CreatedAt    time.Time   `json:"created_at"`
 	UpdatedAt    time.Time   `json:"updated_at"`
 
@@ -30,11 +30,11 @@ type Order struct {
 	SalespersonID   *uuid.UUID `json:"salesperson_id,omitempty"`
 	SalespersonName string     `json:"salesperson_name,omitempty"`
 
-	// Margin & Commission (computed from line-level product cost data)
-	TotalCost       float64 `json:"total_cost"`
-	TotalMargin     float64 `json:"total_margin"`
-	MarginPercent   float64 `json:"margin_percent"`
-	TotalCommission float64 `json:"total_commission"`
+	// Margin & Commission (computed from line-level product cost data, all in cents)
+	TotalCost       int64   `json:"total_cost"`       // Cents
+	TotalMargin     int64   `json:"total_margin"`     // Cents
+	MarginPercent   float64 `json:"margin_percent"`   // Percentage (kept as float)
+	TotalCommission int64   `json:"total_commission"` // Cents
 
 	// Relations
 	Lines []OrderLine `json:"lines,omitempty"`
@@ -46,13 +46,13 @@ type OrderLine struct {
 	ProductID        uuid.UUID  `json:"product_id"`
 	ProductSKU       string     `json:"product_sku,omitempty"`
 	ProductName      string     `json:"product_name,omitempty"`
-	Quantity         float64    `json:"quantity"`
-	PriceEach        float64    `json:"price_each"`
-	UnitCost         float64    `json:"unit_cost"`
-	CommissionRate   float64    `json:"commission_rate"`
+	Quantity         float64    `json:"quantity"`          // Physical quantity (not money)
+	PriceEach        int64      `json:"price_each"`        // Cents
+	UnitCost         int64      `json:"unit_cost"`         // Cents
+	CommissionRate   float64    `json:"commission_rate"`   // Percentage (kept as float)
 	IsSpecialOrder   bool       `json:"is_special_order"`
 	VendorID         *uuid.UUID `json:"vendor_id,omitempty"`
-	SpecialOrderCost float64    `json:"special_order_cost,omitempty"`
+	SpecialOrderCost int64      `json:"special_order_cost,omitempty"` // Cents
 }
 
 type CreateOrderRequest struct {
@@ -63,11 +63,11 @@ type CreateOrderRequest struct {
 
 type OrderLineRequest struct {
 	ProductID        uuid.UUID  `json:"product_id"`
-	Quantity         float64    `json:"quantity"`
-	PriceEach        float64    `json:"price_each"`
+	Quantity         float64    `json:"quantity"`          // Physical quantity (not money)
+	PriceEach        int64      `json:"price_each"`        // Cents
 	IsSpecialOrder   bool       `json:"is_special_order"`
 	VendorID         *uuid.UUID `json:"vendor_id"`
-	SpecialOrderCost float64    `json:"special_order_cost"`
+	SpecialOrderCost int64      `json:"special_order_cost"` // Cents
 }
 
 type UpdateStatusRequest struct {
